Moves 'smile' and 'frown' variable to Preset.js file
