+++ life-game/src/App.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import Grid from './components/grid/Grid';
-// import Preset from './components/preset/Preset';
 import Rules from './components/rules/Rules';
 import About from './components/about/About';
 
@@ -11,7 +10,6 @@ function App() {
     <div className='App'>
       <h1>Conway's Game of Life</h1>
       <Grid />
-      {/* <Preset /> */}
       <Rules />
       <About />
     </div>
